requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2020.6.20
charset-normalizer==2.0.10
idna==3.3
oauthlib==3.1.1
pyairtable==1.0.0.post1
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.0
schedule==1.1.0
six==1.16.0
tornado==6.1
tweepy==4.4.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
